feature: adding travis-ci validation for podcasts.json
A few days ago, I visited this repo site and a problem was ocurring. During the investigation I found a sintax error on podcasts.json: unexpected comma. So I opened a PR fixing this.

But this is basic error, so I think will be great if we have an automate way to garantee that problems like that doesn't ocurr anymore. With Travis-CI we can do that.